+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 - support kubernetes
 - the unique volume of host to container in server and bot
 - try hard-linking when move file from local server to bot-output in local-server mode
+- use `data` instead of `output` as the argument name
+- `-f` for favorite, `-F` for dislike
 
 ## [0.4.0] - 2025-03-12